% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{calc_screen_diam}
\alias{calc_screen_diam}
\title{Calculates screen diameter}
\usage{
calc_screen_diam(alpha, screen_distance)
}
\description{
alpha = alpha angles
screen_distance = distance simulation will be viewed at
}
\keyword{internal}
