% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_speed_model.R
\name{constant_speed_model}
\alias{constant_speed_model}
\title{Constant Speed Model}
\usage{
constant_speed_model(...)
}
\arguments{
\item{screen_distance}{numeric. Distance (cm) from the playback screen to your specimen.}

\item{anim_frame_rate}{numeric. Frames per second (Hz) you want the resulting animation to be.}

\item{speed}{numeric. Speed (cm/s) of the hypothetical approaching attacker.}

\item{attacker_diameter}{numeric. Diameter of the hypothetical approaching attacker}

\item{start_distance}{numeric. Starting distance of the hypothetical approaching attacker}
}
\value{
List object containing the input parameters and the resulting diameter for each
 frame in the animation.
}
\description{
\code{constant_speed_model} creates the screen diameters for a modelled object
  approaching at a constant speed
}
\details{
Creates the screen diameters for a modelled object of specified size approaching at a constant speed,
from a specified distance away. The output list object can be used to create a looming animation in
\code{\link{looming_animation}}. An object screen diameter is calculated for each frame from the
specified starting distance until the hypothetical distance between the attacker and target is zero.

Requires distance from the screen at which the observing specimen will be located, and the intended
frame rate at which the subsequent animation will be played. Note: if the specimen is closer or further
away than the specified screen distance, the animation will be perceived as a different distance and a
different velocity. These details are important in experiments where you want to precisely determine
at what time, perceived speed, or perceived velocity of an attack an escape response may occur at.

If you need to create a looming animation simply to elicit a response, and are not concerned with the
precise details, the function is more than capable of this. Simply enter some inputs, use
\code{\link{looming_animation}} to create the animation, play it back and if it is not what you want,
go back and vary the inputs until you are happy with the result.

Inputs should be in \code{cm}, speed in \code{cm/s}, and frame rate in \code{Hz} or
\code{Frames per Second}.
}
\examples{
loom_model <- constant_speed_model(
                     screen_distance = 20,
                     anim_frame_rate = 60,
                     speed = 500,
                     attacker_diameter = 50,
                     start_distance = 1000)

}
\seealso{
\code{\link{looming_animation}}, \code{\link{looming_animation_calib}}
}
\author{
Nicholas Carey - \link{nicholascarey@gmail.com}
}
