% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{image_progress}
\alias{image_progress}
\title{Display image generation progress bar}
\usage{
image_progress(x, max = 100)
}
\description{
Display image generation progress bar
}
\details{
Displays an updating progress bar within the frame image generation loop. Modified from
\url{https://github.com/klmr/rcane/blob/master/system.R}.
}
\keyword{internal}
