% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loomeR-package.R
\docType{package}
\name{loomeR-package}
\alias{loomeR}
\alias{loomeR-package}
\title{loomeR: Looming Animations for Behavioural and Neurological Experiments}
\description{
loomeR is an R package for creating looming animations for use in behavioural 
  and neurological experiments. In addition, escape response parameters, for example the Apparent 
  Looming Threshold (ALT) of an escape, can be determined. Looming animations are used in a range of 
  physiological, psychological and behavioural sciences to simulate an approaching 
  threat and investigate phenomena such as perception, visual latency, escape responses, 
  and neurological functioning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicholascarey/loomeR}
  \item \url{https://doi.org/10.5281/zenodo.1212570}
  \item Report bugs at \url{https://github.com/nicholascarey/loomeR/issues}
}

}
\keyword{internal}
