% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looming_animation.R,
%   R/looming_animation_calib.R
\name{os}
\alias{os}
\title{Check operating system}
\usage{
os()

os()
}
\description{
This is an internal function.

This is an internal function.
}
\keyword{internal}
