% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{prep_progress}
\alias{prep_progress}
\title{Display preparing video progress bar}
\usage{
prep_progress(x, max = 100)
}
\description{
Display preparing video progress bar
}
\details{
Simply implements a delay
}
\keyword{internal}
