% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{vect_rm_end}
\alias{vect_rm_end}
\title{Remove last n values from a vector}
\usage{
vect_rm_end(x, n = 1)
}
\description{
Remove last n values from a vector
}
\details{
Removes the specified nunmber of values/entries from the end of a vector
}
\keyword{internal}
