% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diameter_model.R
\name{diameter_model}
\alias{diameter_model}
\title{Diameter Model}
\usage{
diameter_model(start_diameter = 5, end_diameter = 50, duration = 5,
  anim_frame_rate = 60, expansion = "constant")
}
\arguments{
\item{start_diameter}{numeric. Diameter (cm) you want the animation to start
at.}

\item{end_diameter}{numeric. Diameter (cm) you want the animation to end at.
Should be slightly larger than the screen size if you want it to fill the
screen in the final frames.}

\item{duration}{numeric. Total duraion (s) you want the animation to be.}

\item{anim_frame_rate}{numeric. Frames per second (Hz) you want the resulting
animation to be.}

\item{expansion}{string. Sets if the expansion of the simulation is modelled
as a constant speed or constant increase in diameter (see Details).}
}
\value{
List object containing the input parameters and the resulting
  diameter for each frame in the animation.
}
\description{
Creates a simple model for making a looming animation by setting
  starting screen diameter and total duration
}
\details{
Exapnsion of the simulation can be set in two ways. If
  \code{constant}, the function models a constant speed of the attacker.
  Because of visual foreshortening, this results in a simulation that expands
  progressively more rapidly towards the end of the animation. If \code{=
  diameter}, the function instead imposes a constant increase in diameter,
  i.e. a simulation that expands by the same amount in diameter in each
  frame. This simulation represents an oncoming object that actually slows
  down as it gets closer to the target.

  Inputs should be in \code{cm}, speed in \code{cm/s}, and frame rate in
  \code{Hz} or \code{Frames per Second}.
}
\examples{
loom_model <- diameter_model(
                     start_diameter = 5,
                     end_diameter = 50,
                     duration = 5,
                     anim_frame_rate = 60,
                     expansion = "constant")

}
\seealso{
\code{\link{looming_animation}},
  \code{\link{looming_animation_calib}},  \code{\link{variable_speed_model}}
}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
