% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alt.R
\name{get_alt}
\alias{get_alt}
\title{Determine the Apparent Looming Threshold (ALT) for a particular
  response frame}
\usage{
get_alt(x, response_frame = NULL, new_distance = NULL, latency = 0)
}
\arguments{
\item{x}{numeric. Object of class}

\item{response_frame}{integer. The frame of the animation at which the
subject responds, or the frame from which you want to extract the ALT and
other parameters.}

\item{new_distance}{numeric. The distance in cm the specimen is from screen,
if this is different to that used to create the model.}

\item{latency}{numeric. Visual response latency in seconds.}
}
\value{
List object of class \code{get_alt}
}
\description{
\code{get_alt} gets the Apparent Looming Threshold (ALT),
  the rate of change of the visual angle of the looming shape (da/dt in
  radians/s) at a particlar frame. Typically this is the frame a specimen
  responds to the stimulus, or the response taking into account a delay due
  to response latency. It also determines the perceived speed and perceived
  distance of the looming shape at this frame. It also allows the extracted
  parameters to be 'corrected' in two ways. If the observing specimen has
  moved, and so is at a different distance from the screen as specified in
  the original model, it will perceive a different distance, speed and ALT.
  This function allows the perceived parameters to be adjusted by specifying
  a \code{new_distance}. The function also allows a response latency to be
  applied (\code{latency}). Typically, there is a lag between an animal
  responding or "deciding" to respond to a stimulus, and this signal reaching
  their musculature.
}
\details{
Gets the ALT, perceived speed and perceived distance of a specimen

  Since da/dt is a derivative, its value lies between frames. The function
  uses the value between the response frame and the next.

  Latency: Latency is applied internally by adjusting the response frame. The
  frame rate multiplied by the latency gives the number of frames 'back' from
  the entered \code{response_frame} the function goes to extract the
  parameters. In the event the number of frames backwards is a decimal, this
  is rounded to the nearest integer. For example, with 60 fps video and 0.06s
  latency, the function would seek to go back 3.6 frames. Instead, this is
  rounded to 4.

  Diameter model: Note that for a diameter model, the dadt/ALT can be
  extracted (if a viewing distance is supplied), but *not* perceived distance
  and speed. This is because the size of the attacker was never specified in
  creating the model. While the viewing angle (alpha) can be calculated, and
  thus the da/dt, the perceived distance and/or speed cannot. A small object
  on screen expanding rapidly could represent a very small object moving
  slowly at a close distance, or a very large object moving rapidly at a far
  distance. Both of these can produce an identical da/dt, but without a size
  being specified the speed and distance cannot be determined. Expressing the
  ALT as a da/dt value is designed to negate this issue, by focussing purely
  on the rate of change of the viewing angle. If these parameters may be
  important to your study however, the model should be created in
  \code{\link{constant_speed_model}} or \code{\link{variable_speed_model}}.
  Note, \code{\link{diameter_model}} is intended to produce simple animations
  purely to induce response, where the precise parameters are not important.
}
\examples{
get_alt(x, ...)

}
\seealso{
\code{\link{looming_animation}}, \code{\link{???}}
}
\author{
Nicholas Carey - \link{nicholascarey@gmail.com}
}
