% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alt.R
\name{get_alt}
\alias{get_alt}
\title{Determine the Apparent Looming Threshold (ALT) for a particular
  response frame}
\usage{
get_alt(...)
}
\arguments{
\item{x}{numeric. Object of class}

\item{response_frame}{integer. The frame of the animation at which the
subject responds, or the frame from which you want to extract the ALT and
other parameters.}

\item{new_distance}{numeric. The distance in cm the specimen is from screen,
if this is different to that used to create the model.}

\item{latency}{numeric. Visual response latency in milliseconds.}
}
\value{
List object of class \code{get_alt}
}
\description{
\code{get_alt} gets the Apparent Looming Threshold (ALT) that is
  the rate of change of the visual angle of the looming shape (da/dt in
  radians/s) at a particlar frame. Typically this is the frame a specimen
  responds to the stimulus, or the response taking into account a delay due
  to response latency. It also determines the perceived speed and perceived
  distance of the looming shape at this frame. It also allows the extracted
  parameters to be 'corrected' in two ways. If the observing specimen has
  moved, and so is at a different distance from the screen as specified in
  the original model, it will perceive a different distance, speed and ALT.
  This function allows the perceived parameters to be adjusted by specifying
  a \code{new_distance}. The function also allows a response latency to be
  applied (\code{latency}). Typically, there is a lag between an animal
  responding or "deciding" to respond to a stimulus, and this signal reaching
  their musculature.
}
\details{
Gets the ALT, perceived speed and perceived distance of a specimen

  Since da/dt is a derivative, its value lies between frames. The function
  uses the value between the response frame and the next.
}
\examples{
get_alt(x, ...)

}
\seealso{
\code{\link{looming_animation}}, \code{\link{???}}
}
\author{
Nicholas Carey - \link{nicholascarey@gmail.com}
}
