% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loomeR-package.R
\docType{package}
\name{loomeR-package}
\alias{loomeR}
\alias{loomeR-package}
\title{loomeR: Create a looming animation}
\description{
Creates a model of a object of specified size approaching at specified speed.
Uses that model to then create a looming animation.
}
\keyword{internal}
