% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loomeR-package.R
\docType{package}
\name{loomeR-package}
\alias{loomeR}
\alias{loomeR-package}
\title{loomeR: Create a looming animation}
\description{
Create a looming animation for use in behavioural and neurological experiments.
}
\keyword{internal}
