% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loomeR-package.R
\docType{package}
\name{loomeR-package}
\alias{loomeR}
\alias{loomeR-package}
\title{loomeR: Looming Animations for Use in Behavioural and Neurological Experiments}
\description{
Create and analyse looming animations for use in behavioural and neurological experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicholascarey/loomeR}
  \item \url{https://doi.org/10.5281/zenodo.1212570}
  \item Report bugs at \url{https://github.com/nicholascarey/loomeR/issues}
}

}
\keyword{internal}
